(#22205) Extract common Prioritizer code into a superclass
All of the prioritizers had the same code to track priorities already
produced. This pulls that into a superclass.
